Add ability for a per project .vimrc.local file
I was just thinking about this earlier today and then noticed on Twitter
that @garybernhardt was thinking the same thing. From some of the
replies to his tweet, I came up wtih this. Adding a .vimrc.local file to
the CWD where vim is opened will cause it to be source, meaning custom
key bindings can be applied on a per project basis.