feat(vim): better repeat movement support
For normal mode overrides like j -> gj and k -> gk, only perform
override if v:count is 0. Otherwise, run the normal command. The reason
for this check is to check if there is a higher number for repeat
commands, such as expecting the . (dot) command repeat to work on both
lines with wrapping and those without.