feat(tmux): default window name shown to path
If no window name is manually set, either when the window is created or
later with rename-window (prefix-,), default the window name to the
current path.

This is accomplished by enabling automatic-rename and setting an
automatic-rename-format to "" so that there exists no window name, which
is then used in the ternary for the window name format (#{?}).